new variables (uflx, dflx, edel), expose monoenergetic, and typos
Exposed the monoenergetic function, included 3 new variables dflx, uflx, edel (useful for electron distribution calculations), and fixed a typo in doc (keV to eV). I also changed the implementation of monoenergetic to set echar=emono/2., such that the tiny additional flux is added at the same energy as the primary beam. Previously, the energy was set very low, which led to a huge number flux in the low-energy bins.